Crate provider

A module to help parse command line arguments for flow URLs and fetch the associated content

Modules

args
Help take file/url strings from a command line and convert them into URLs (as Strings) with schemes for use with flowlibc and flowlibr.

content
Content provider trait. It defines methods for for getting content of flows from files, http or library references.

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.