Struct provider::lib_provider::MetaProvider

pub struct MetaProvider { /* fields omitted */ }

The MetaProvider implements the Provider trait and based on the url and it’s resolution to a real location for content invokes one of the child providers it has to fetch the content (e.g. File or Http).

Implementations

impl MetaProvider

Instantiate MetaProvider and then use the Provider trait methods on it to resolve and fetch content depending on the URL scheme.

use simpath::Simpath;
use url::Url;
use provider::lib_provider::{LibProvider, MetaProvider};
let lib_search_path = Simpath::new_with_separator("FLOW_LIB_PATH", ',');
let meta_provider = &MetaProvider::new(lib_search_path) as &dyn LibProvider;
let url = Url::parse("file://directory").unwrap();
match meta_provider.resolve_url(&url, "default", &["toml"]) {
    Ok((resolved_url, lib_ref)) => {
        match meta_provider.get_contents(&resolved_url) {
            Ok(contents) => println!("Content: {:?}", contents),
            Err(e) => println!("Got error '{}'", e)
        }
    }
    Err(e) => {
        println!("Found error '{}'", e);
    }
};

pub fn new(lib_search_path: Simpath) -> Self

Create a new MetaProvider initializing it with a search path for libraries