Struct provider::NativeProvider

pub struct NativeProvider<T: FileSystem + Send + Sync> { /* private fields */ }

Implementations

impl<T: FileSystem + Send + Sync> NativeProvider<T>

pub fn new( namespace: impl Into<String>, config_path: impl Into<String>, tmp_dir: impl Into<String>, filesystem: T ) -> Self

Zombienet native provider allows to run the nodes as a local process in the local environment params: namespace: Namespace of the clien config_path: Path where configuration relies tmp_dir: Temporary directory where files will be placed filesystem: Filesystem to use (std::fs::FileSystem, mock etc.)